refactor(ip): migrate dashboard.order-legacy to component
