+++ packages/manager/apps/dedicated/client/app/ip/dashboard/order-legacy/DEDICATED.html
