+++ packages/manager/apps/dedicated/client/app/ip/dashboard/order-legacy/PCC.html
@@ -71,7 +71,7 @@
     <oui-field
         data-label="{{:: 'ip_order_estimatedclientsnumber' | translate }}"
         label-popover="{{:: 'ip_order_estimatedclientsnumber_help' | translate }}"
-        data-ng-if="model.service.usesLegacyOrder"
+        data-ng-if="model.service.usesOrderLegacy"
     >
         <input
             type="number"
