+++ packages/manager/apps/dedicated/client/app/ip/dashboard/order-legacy/VPS.html
