+++ packages/manager/apps/dedicated/client/app/ip/dashboard/order-legacy/order-legacy.component.js
@@ -1,0 +1,11 @@
+import controller from './order-legacy.controller';
+import template from './order-legacy.html';
+
+export default {
+  bindings: {
+    goBack: '<',
+    goToOrganisation: '<',
+  },
+  controller,
+  template,
+};
