+++ packages/manager/apps/dedicated/client/app/ip/dashboard/order-legacy/order-legacy.routing.js
@@ -1,0 +1,18 @@
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider.state('app.ip.dashboard.order-legacy', {
+    url: '/orderLegacy',
+    resolve: {
+      goBack: /* @ngInject */ (goToDashboard) => goToDashboard,
+      goToOrganisation: /* @ngInject */ ($state) => (
+        params,
+        transitionParams,
+      ) => $state.go('app.ip.organisation', params, transitionParams),
+    },
+    views: {
+      modal: {
+        component: 'ipDashboardOrderLegacy',
+      },
+    },
+    layout: 'modal',
+  });
+};
